LayoutTests/media/media-source/media-managedmse-audible-setrate-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
EVENT(update)
EVENT(update)
EVENT(update)
EVENT(update)
EVENT(update)
RUN(source.endOfStream())
EVENT(sourceended)
RUN(video.currentTime=3)
EVENT(seeked)
RUN(video.play())
EVENT(playing)
EVENT(timeupdate)
Audio check: silence (0% audible)
RUN(video.playbackRate = 0.5)
EVENT(timeupdate)
Audio check: silence (0% audible)
END OF TEST
